pyproject.toml
[tool.poetry]
name = "idloader"
version = "0.1.0"
description = "A mod loader and managemer for Quake"
authors = ["ranguli <hello@joshmurphy.ca>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
xmltodict = "^0.12.0"
wheel = "^0.34.2"
PyQt = "5.14.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
sphinx = "^2.4.4"
sphinx_rtd_theme = "^0.4.3"
pytest = "^5.4.1"
tox = "^3.14.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
